fix(token): Integers representation changed from usize to i64
Because we probably want to have negative integers.

+++ src/ast.rs
@@ -72,7 +72,7 @@ impl Display for ExpressionStatement {
 #[derive(Debug)]
 pub enum Expression {
     Identifier(Identifier),
-    IntegerLiteral(usize),
+    IntegerLiteral(i64),
 }
 
 impl Display for Expression {
